+++ NEWS.md
@@ -4,6 +4,9 @@
 to choose whether the path approximating the ellipse will be interior or 
 exterior (thanks to Duncan Murdoch for the help).
 
+- New method `perimeter` for the `Ellipse` class; it returns the perimeter 
+of the ellipse.
+
 
 # PlaneGeometry 1.4.0 (2023-05-19)